feat(condo): Added some tests
