+++ apps/condo/domains/resident/schema/RegisterServiceConsumerService.js
@@ -279,4 +279,5 @@ const RegisterServiceConsumerService = new GQLCustomSchema('RegisterServiceConsu
 
 module.exports = {
     RegisterServiceConsumerService,
+    ERRORS,
 }
