+++ apps/condo/domains/resident/schema/RegisterServiceConsumerService.test.js
@@ -3,784 +3,130 @@
  */
 
 const { faker } = require('@faker-js/faker')
-const { fetch }  = require('@open-condo/keystone/fetch')
 
-const { makeClient, expectToThrowGQLError } = require('@open-condo/keystone/test.utils')
-const { makeLoggedInAdminClient } = require('@open-condo/keystone/test.utils')
+const { fetch }  = require('@open-condo/keystone/fetch')
 const {
-    catchErrorFrom,
     expectToThrowAccessDeniedErrorToObj,
     expectToThrowAuthenticationErrorToObj,
-    expectToThrowAuthenticationErrorToObjects,
+    expectToThrowGQLError,
 } = require('@open-condo/keystone/test.utils')
 
 const { CONTEXT_FINISHED_STATUS, CONTEXT_IN_PROGRESS_STATUS } = require('@condo/domains/acquiring/constants/context')
-const { addAcquiringIntegrationAndContext } = require('@condo/domains/acquiring/utils/testSchema')
 const { AcquiringTestMixin } = require('@condo/domains/acquiring/utils/testSchema/acquiringTestMixin')
 const { 
-    ONLINE_INTERACTION_CHECK_ACCOUNT_ERROR_STATUS, 
-    ONLINE_INTERACTION_CHECK_ACCOUNT_SUCCESS_STATUS, 
+    ONLINE_INTERACTION_CHECK_ACCOUNT_SUCCESS_STATUS,
     ONLINE_INTERACTION_CHECK_ACCOUNT_NOT_FOUND_STATUS, 
 } = require('@condo/domains/billing/utils/serverSchema/checkBillingAccountNumberForOrganization')
-const { addBillingIntegrationAndContext } = require('@condo/domains/billing/utils/testSchema')
-const {
-    createTestBillingProperty,
-    createTestBillingAccount,
-    createTestBillingIntegration,
-    createTestBillingIntegrationOrganizationContext,
-    makeContextWithOrganizationAndIntegrationAsAdmin,
-    updateTestBillingIntegrationOrganizationContext,
-} = require('@condo/domains/billing/utils/testSchema')
 const { BillingTestUtils } = require('@condo/domains/billing/utils/testSchema/utils')
-const { NOT_FOUND } = require('@condo/domains/common/constants/errors')
-const {
-    COLD_WATER_METER_RESOURCE_ID,
-    HOT_WATER_METER_RESOURCE_ID,
-} = require('@condo/domains/meter/constants/constants')
-const { MeterResource, createTestMeter, Meter } = require('@condo/domains/meter/utils/testSchema')
-const { SERVICE_PROVIDER_TYPE } = require('@condo/domains/organization/constants/common')
-const { createTestOrganization } = require('@condo/domains/organization/utils/testSchema')
-const {
-    createTestProperty,
-    makeClientWithProperty,
-} = require('@condo/domains/property/utils/testSchema')
+const { MeterTestMixin } = require('@condo/domains/meter/utils/testSchema/meterTestMixin')
+const { PropertyTestMixin } = require('@condo/domains/property/utils/testSchema/propertyTestMixin')
+const { ERRORS: { BILLING_ACCOUNT_NOT_FOUND, ACCOUNT_NUMBER_IS_NOT_SPECIFIED } } = require('@condo/domains/resident/schema/RegisterServiceConsumerService')
 const {
     registerServiceConsumerByTestClient,
-    updateTestServiceConsumer,
-    createTestResident,
-    updateTestResident,
-    registerResidentServiceConsumersByTestClient,
 } = require('@condo/domains/resident/utils/testSchema')
 const { ResidentTestMixin } = require('@condo/domains/resident/utils/testSchema/residentTestMixin')
-const { RESIDENT } = require('@condo/domains/user/constants/common')
-const { updateTestUser, makeClientWithResidentUser } = require('@condo/domains/user/utils/testSchema')
 
-
-describe('RegisterResidentServiceConsumers', () => {
-    let adminClient
-    let residentClient
-    let anonymousClient
-
-    let organization
-
-    beforeAll(async () => {
-        adminClient = await makeLoggedInAdminClient()
-        residentClient = await makeClientWithResidentUser()
-        anonymousClient = await makeClient()
-
-        organization = (await createTestOrganization(adminClient))[0]
-    })
-
-    describe('Admin', () => {
-        it('can register ServiceConsumer via this mutation', async () => {
-            const [property] = await createTestProperty(adminClient, organization)
-            const unitName = faker.random.alphaNumeric(8)
-            const accountNumber = faker.random.alphaNumeric(8)
-
-            const [resident] = await createTestResident(adminClient, residentClient.user, property, {
-                unitName,
-            })
-
-            const [resource] = await MeterResource.getAll(adminClient, { id: COLD_WATER_METER_RESOURCE_ID })
-            const [meter] = await createTestMeter(adminClient, organization, property, resource, {
-                accountNumber, unitName,
-            })
-
-            const [data] = await registerResidentServiceConsumersByTestClient(adminClient, {
-                resident: { id: resident.id },
-                accountNumber,
-            })
-
-            expect(data).toHaveLength(1)
-            expect(data[0]).toHaveProperty(['organization', 'id'], organization.id)
-            expect(data[0]).toHaveProperty(['resident', 'id'], resident.id)
-            expect(data[0]).toHaveProperty('accountNumber', meter.accountNumber)
-        })
-    })
-    describe('Resident', () => {
-        it('can register ServiceConsumer if organization has connected billing', async () => {
-            const { organization, context } = await makeContextWithOrganizationAndIntegrationAsAdmin()
-            await updateTestBillingIntegrationOrganizationContext(adminClient, context.id, { status: CONTEXT_FINISHED_STATUS })
-            const [property] = await createTestProperty(adminClient, organization)
-            const [billingProperty] = await createTestBillingProperty(adminClient, context, {
-                address: property.address,
-            })
-            const [billingAccount] = await createTestBillingAccount(adminClient, context, billingProperty)
-            const {
-                acquiringIntegration,
-                acquiringIntegrationContext,
-            } = await addAcquiringIntegrationAndContext(adminClient, organization, {}, {
-                status: CONTEXT_FINISHED_STATUS,
-            })
-
-            const [resident] = await createTestResident(adminClient, residentClient.user, property, {
-                unitName: billingAccount.unitName,
-            })
-
-            const [data] = await registerResidentServiceConsumersByTestClient(residentClient, {
-                resident: { id: resident.id },
-                accountNumber: billingAccount.number,
-            })
-
-            expect(data).toHaveLength(1)
-            expect(data[0]).toHaveProperty(['organization', 'id'], organization.id)
-            expect(data[0]).toHaveProperty('accountNumber', billingAccount.number)
-            expect(data[0]).toHaveProperty(['resident', 'id'], resident.id)
-            expect(data[0]).toHaveProperty(['residentAcquiringIntegrationContext', 'id'], acquiringIntegrationContext.id)
-            expect(data[0]).toHaveProperty(['residentAcquiringIntegrationContext', 'integration', 'id'], acquiringIntegration.id)
-            expect(data[0]).toHaveProperty(['residentAcquiringIntegrationContext', 'integration', 'hostUrl'], acquiringIntegration.hostUrl)
-        })
-
-        it('doesn\'t create same ServiceConsumer twice', async () => {
-            const { organization, context } = await makeContextWithOrganizationAndIntegrationAsAdmin()
-            await updateTestBillingIntegrationOrganizationContext(adminClient, context.id, { status: CONTEXT_FINISHED_STATUS })
-            const [property] = await createTestProperty(adminClient, organization)
-            const [billingProperty] = await createTestBillingProperty(adminClient, context, {
-                address: property.address,
-            })
-            const [billingAccount] = await createTestBillingAccount(adminClient, context, billingProperty)
-            const [resident] = await createTestResident(adminClient, residentClient.user, property, {
-                unitName: billingAccount.unitName,
-            })
-
-            const [data] = await registerResidentServiceConsumersByTestClient(residentClient, {
-                resident: { id: resident.id },
-                accountNumber: billingAccount.number,
-            })
-
-            expect(data).toHaveLength(1)
-
-            const [newData] = await registerResidentServiceConsumersByTestClient(residentClient, {
-                resident: { id: resident.id },
-                accountNumber: billingAccount.number,
-            })
-
-            expect(newData).toHaveLength(1)
-            expect(newData[0]).toHaveProperty('id', data[0].id)
-        })
-
-        it('can restore existing ServiceConsumer after softDelete operation', async () => {
-            const { organization, context } = await makeContextWithOrganizationAndIntegrationAsAdmin()
-            await updateTestBillingIntegrationOrganizationContext(adminClient, context.id, { status: CONTEXT_FINISHED_STATUS })
-            const [property] = await createTestProperty(adminClient, organization)
-            const [billingProperty] = await createTestBillingProperty(adminClient, context, {
-                address: property.address,
-            })
-            const [billingAccount] = await createTestBillingAccount(adminClient, context, billingProperty)
-
-            const [resident] = await createTestResident(adminClient, residentClient.user, property, {
-                unitName: billingAccount.unitName,
-            })
-
-            const [data] = await registerResidentServiceConsumersByTestClient(residentClient, {
-                resident: { id: resident.id },
-                accountNumber: billingAccount.number,
-            })
-
-            expect(data).toHaveLength(1)
-            expect(data[0]).toHaveProperty(['resident', 'id'], resident.id)
-            expect(data[0]).toHaveProperty(['organization', 'id'], organization.id)
-            expect(data[0]).toHaveProperty('accountNumber', billingAccount.number)
-
-            await updateTestServiceConsumer(residentClient, data[0].id, { deletedAt: data[0].createdAt })
-
-            const [restoredData] = await registerResidentServiceConsumersByTestClient(residentClient, {
-                resident: { id: resident.id },
-                accountNumber: billingAccount.number,
-            })
-
-            expect(restoredData).toHaveLength(1)
-            expect(restoredData[0]).toHaveProperty('id', data[0].id)
-            expect(restoredData[0]).toHaveProperty(['resident', 'id'], data[0].resident.id)
-            expect(restoredData[0]).toHaveProperty('accountNumber', data[0].accountNumber)
-        })
-
-        it('creates ServiceConsumer if organization has Meter with same accountNumber', async () => {
-            const [property] = await createTestProperty(adminClient, organization)
-            const [resource] = await MeterResource.getAll(adminClient, { id: COLD_WATER_METER_RESOURCE_ID })
-            const unitName = faker.random.alphaNumeric(8)
-            const accountNumber = faker.random.alphaNumeric(8)
-            const [resident] = await createTestResident(adminClient, residentClient.user, property, {
-                unitName,
-            })
-            const [meter] = await createTestMeter(adminClient, organization, property, resource, {
-                unitName, accountNumber,
-            })
-
-            const [data] = await registerResidentServiceConsumersByTestClient(residentClient, {
-                resident: { id: resident.id },
-                accountNumber,
-            })
-
-            expect(data).toHaveLength(1)
-            expect(data[0]).toHaveProperty(['resident', 'id'], resident.id)
-            expect(data[0]).toHaveProperty(['organization', 'id'], organization.id)
-            expect(data[0]).toHaveProperty('accountNumber', meter.accountNumber)
-        })
-
-        it('creates 2 ServiceConsumers for 2 organizations based on resident address equality', async () => {
-            const [serviceProviderOrganization] = await createTestOrganization(adminClient, {
-                type: SERVICE_PROVIDER_TYPE,
-            })
-            const [property] = await createTestProperty(adminClient, organization)
-            const [serviceProviderProperty] = await createTestProperty(adminClient, serviceProviderOrganization, {
-                address: property.address,
-            })
-            const unitName = faker.random.alphaNumeric(8)
-            const accountNumber = faker.random.alphaNumeric(8)
-
-            // Resident can't be created to organization with type = service-provider
-            const [resident] = await createTestResident(adminClient, residentClient.user, property, {
-                unitName,
-            })
-
-            const [coldResource] = await MeterResource.getAll(adminClient, { id: COLD_WATER_METER_RESOURCE_ID })
-            const [hotResource] = await MeterResource.getAll(adminClient, { id: HOT_WATER_METER_RESOURCE_ID })
-
-            // Same accountNumber, address and unitName
-            await createTestMeter(adminClient, organization, property, coldResource, {
-                accountNumber, unitName,
-            })
-            await createTestMeter(adminClient, serviceProviderOrganization, serviceProviderProperty, hotResource, {
-                accountNumber, unitName,
-            })
-
-            const [data] = await registerResidentServiceConsumersByTestClient(residentClient, {
-                resident: { id: resident.id },
-                accountNumber,
-            })
-
-            expect(data).toHaveLength(2)
-            expect(data.find(e => e.organization.id === organization.id)).toBeDefined()
-            expect(data.find(e => e.organization.id === serviceProviderOrganization.id)).toBeDefined()
-        })
-    })
-    describe('Anonymous', () => {
-        it('can\'t execute this mutation', async () => {
-            await expectToThrowAuthenticationErrorToObjects(async () => {
-                await registerResidentServiceConsumersByTestClient(anonymousClient, {
-                    resident: { id: faker.datatype.uuid() },
-                    accountNumber: faker.random.alphaNumeric(8),
-                })
-            })
-        })
-    })
-
-    describe('Validations', () => {
-        it('throw an error when no billingAccount and Meters were found', async () => {
-            const userClient = await makeClientWithProperty()
-
-            const [integration] = await createTestBillingIntegration(adminClient)
-            await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, integration)
-
-            await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-            const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-                unitName: '21',
-            })
-
-            const payload = {
-                resident: { id: resident.id },
-                accountNumber: '221231232',
-            }
-
-            await expectToThrowGQLError(async () => {
-                await registerResidentServiceConsumersByTestClient(userClient, payload)
-            }, {
-                mutation: 'registerResidentServiceConsumers',
-                variable: ['data', 'accountNumber'],
-                code: 'BAD_USER_INPUT',
-                type: NOT_FOUND,
-                message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and address',
-                messageForUser: 'api.resident.RegisterServiceConsumerService.BILLING_ACCOUNT_OR_METER_NOT_FOUND',
-            }, 'objs')
-        })
-
-        it('throw an error when accountNumber is empty', async () => {
-            const userClient = await makeClientWithProperty()
-            await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-            const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-                unitName: faker.random.alphaNumeric(8),
-            })
-
-            const payload = {
-                resident: { id: resident.id },
-                accountNumber: '',
-            }
-
-            await expectToThrowGQLError(async () => {
-                await registerResidentServiceConsumersByTestClient(userClient, payload)
-            }, {
-                mutation: 'registerServiceConsumer',
-                variable: ['data', 'accountNumber'],
-                code: 'BAD_USER_INPUT',
-                type: 'WRONG_FORMAT',
-                message: 'Argument "accountNumber" is null or empty',
-            }, 'objs')
-        })
-
-        it('throw error when try to create ServiceConsumer for deleted resident', async () => {
-            const userClient = await makeClientWithProperty()
-            await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-            const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-                unitName: faker.random.alphaNumeric(8),
-            })
-
-            await updateTestResident(adminClient, resident.id, { deletedAt: resident.createdAt })
-
-            await expectToThrowGQLError(async () => {
-                await registerResidentServiceConsumersByTestClient(userClient, {
-                    resident: { id: resident.id }, accountNumber: faker.random.alphaNumeric(8),
-                })
-            }, {
-                mutation: 'registerServiceConsumer',
-                variable: ['data', 'residentId'],
-                code: 'BAD_USER_INPUT',
-                type: NOT_FOUND,
-                message: 'Cannot find Resident for current user',
-            }, 'objs')
-        })
-
-        it('throw error when try createServiceConsumer with BillingIntegrationOrganizationContext not in finished status', async () => {
-            const { organization, context } = await makeContextWithOrganizationAndIntegrationAsAdmin({}, {}, {}, true)
-            const [property] = await createTestProperty(adminClient, organization)
-            const [billingProperty] = await createTestBillingProperty(adminClient, context, {
-                address: property.address,
-            })
-            const [billingAccount] = await createTestBillingAccount(adminClient, context, billingProperty)
-            await addAcquiringIntegrationAndContext(adminClient, organization, {}, {
-                status: CONTEXT_FINISHED_STATUS,
-            })
-
-            const [resident] = await createTestResident(adminClient, residentClient.user, property, {
-                unitName: billingAccount.unitName,
-            })
-
-            expect(context).not.toEqual(CONTEXT_FINISHED_STATUS)
-
-            await expectToThrowGQLError(async () => {
-                await registerResidentServiceConsumersByTestClient(residentClient, {
-                    resident: { id: resident.id },
-                    accountNumber: billingAccount.number,
-                })
-            }, {
-                mutation: 'registerResidentServiceConsumers',
-                variable: ['data', 'accountNumber'],
-                code: 'BAD_USER_INPUT',
-                type: NOT_FOUND,
-                message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and address',
-                messageForUser: 'api.resident.RegisterServiceConsumerService.BILLING_ACCOUNT_OR_METER_NOT_FOUND',
-            }, 'objs')
-        })
-
-    })
-})
-
-describe('RegisterServiceConsumerService', () => {
-    let adminClient
-    let integration
-
-    beforeAll(async () => {
-        adminClient = await makeLoggedInAdminClient()
-        integration = (await createTestBillingIntegration(adminClient))[0]
-    })
-
-    it('does not create same service consumer twice', async () => {
-        const userClient = await makeClientWithProperty()
-        const adminClient = await makeLoggedInAdminClient()
-
-        const [context] = await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, integration, { status: CONTEXT_FINISHED_STATUS })
-        const [billingProperty] = await createTestBillingProperty(adminClient, context)
-        const [billingAccountAttrs] = await createTestBillingAccount(adminClient, context, billingProperty)
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: billingAccountAttrs.unitName,
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: billingAccountAttrs.number,
-            organizationId: userClient.organization.id,
-        }
-
-        const out = await registerServiceConsumerByTestClient(userClient, payload)
-        expect(out).not.toEqual(undefined)
-
-        const out2 = await registerServiceConsumerByTestClient(userClient, payload)
-        expect(out2.id).toEqual(out.id)
-    })
-
-    it('allows to create service consumers with same resident and accountNumber for multiple organizations', async () => {
-        const userClient = await makeClientWithProperty()
-
-        const USER_UNIT_NAME = String(faker.datatype.number())
-        const USER_ACCOUNT_NUMBER = String(faker.datatype.number())
-
-        // Org 1 = management company organization
-
-        const organization1 = userClient.organization
-
-        const [resource] = await MeterResource.getAll(adminClient, { id: COLD_WATER_METER_RESOURCE_ID })
-
-
-        await createTestMeter(adminClient, organization1, userClient.property, resource, {
-            unitName: USER_UNIT_NAME,
-            accountNumber: USER_ACCOUNT_NUMBER,
-        })
-        await createTestBillingIntegrationOrganizationContext(adminClient, organization1, integration, { status: CONTEXT_FINISHED_STATUS })
-
-        // Org 2 = just some other org, which provides some services for this resident (like intercom service)
-
-        const [organization2] = await createTestOrganization(adminClient, { type: SERVICE_PROVIDER_TYPE })
-
-        const [context] = await createTestBillingIntegrationOrganizationContext(adminClient, organization2, integration, { status: CONTEXT_FINISHED_STATUS })
-        const [billingProperty] = await createTestBillingProperty(adminClient, context)
-        const [billingAccountAttrs] = await createTestBillingAccount(adminClient, context, billingProperty, {
-            number: USER_ACCOUNT_NUMBER,
-            unitName: USER_UNIT_NAME,
-        })
-
-        // Prepare resident and create serviceConsumers
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: billingAccountAttrs.unitName,
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: billingAccountAttrs.number,
-            organizationId: organization2.id,
-        }
-
-        const [out] = await registerServiceConsumerByTestClient(userClient, payload)
-        expect(out.id).toBeDefined()
-
-        const payload2 = {
-            residentId: resident.id,
-            accountNumber: billingAccountAttrs.number,
-            organizationId: organization1.id,
-        }
-
-        const [out2] = await registerServiceConsumerByTestClient(userClient, payload2)
-        expect(out2.id).toBeDefined()
-
-        expect(out.id === out2.id).toBeFalsy()
-    })
-
-    it('can create, delete and create service consumer', async () => {
-        const userClient = await makeClientWithProperty()
-
-        const [context] = await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, integration, { status: CONTEXT_FINISHED_STATUS })
-        const [billingProperty] = await createTestBillingProperty(adminClient, context)
-        const [billingAccountAttrs] = await createTestBillingAccount(adminClient, context, billingProperty)
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: billingAccountAttrs.unitName,
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: billingAccountAttrs.number,
-            organizationId: userClient.organization.id,
-        }
-        const [out] = await registerServiceConsumerByTestClient(userClient, payload)
-        expect(out).not.toEqual(undefined)
-
-        await updateTestServiceConsumer(userClient, out.id, { deletedAt: 'true' })
-
-        const [out2] = await registerServiceConsumerByTestClient(userClient, payload)
-        expect(out2.id).toEqual(out.id)
-    })
-
-    it('creates serviceConsumer with billingAccount for separate organization', async () => {
-        const userClient = await makeClientWithProperty()
-
-        const [ organization ] = await createTestOrganization(adminClient)
-
-        const { billingIntegrationContext } = await addBillingIntegrationAndContext(adminClient, organization, {}, { status: CONTEXT_FINISHED_STATUS })
-
-        const [billingProperty] = await createTestBillingProperty(adminClient, billingIntegrationContext)
-        const [billingAccountAttrs] = await createTestBillingAccount(adminClient, billingIntegrationContext, billingProperty)
-
-        const { acquiringIntegration, acquiringIntegrationContext } = await addAcquiringIntegrationAndContext(adminClient, organization, {}, { status: CONTEXT_FINISHED_STATUS })
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: billingAccountAttrs.unitName,
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: billingAccountAttrs.number,
-            organizationId: organization.id,
-        }
-
-        const [ out ] = await registerServiceConsumerByTestClient(userClient, payload)
-        expect(out).toBeDefined()
-        expect(out.accountNumber).toEqual(payload.accountNumber)
-        expect(out.resident.id).toEqual(payload.residentId)
-        expect(out.organization.id).toEqual(payload.organizationId)
-        expect(out.residentAcquiringIntegrationContext.id).toEqual(acquiringIntegrationContext.id)
-        expect(out.residentAcquiringIntegrationContext.integration).toBeDefined()
-        expect(out.residentAcquiringIntegrationContext.integration.id).toEqual(acquiringIntegration.id)
-        expect(out.residentAcquiringIntegrationContext.integration.hostUrl).toEqual(acquiringIntegration.hostUrl)
-    })
-
-    it('creates serviceConsumer with billingAccount and Meters', async () => {
-        const userClient = await makeClientWithProperty()
-
-        const USER_UNIT_NAME = String(faker.datatype.number())
-        const USER_ACCOUNT_NUMBER = String(faker.datatype.number())
-
-        const { billingIntegration, billingIntegrationContext } = await addBillingIntegrationAndContext(adminClient, userClient.organization)
-        const [billingProperty] = await createTestBillingProperty(adminClient, billingIntegrationContext)
-        await createTestBillingAccount(adminClient, billingIntegrationContext, billingProperty, {
-            number: USER_ACCOUNT_NUMBER,
-            unitName: USER_UNIT_NAME,
-        })
-
-        await addAcquiringIntegrationAndContext(adminClient, userClient.organization)
-
-        const [resource] = await MeterResource.getAll(adminClient, { id: COLD_WATER_METER_RESOURCE_ID })
-        await createTestMeter(adminClient, userClient.organization, userClient.property, resource, {
-            unitName: USER_UNIT_NAME,
-            accountNumber: USER_ACCOUNT_NUMBER,
-        })
-        await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, billingIntegration, { status: CONTEXT_FINISHED_STATUS })
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: USER_UNIT_NAME,
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: USER_ACCOUNT_NUMBER,
-            organizationId: userClient.organization.id,
-            extra: { paymentCategory: 'Housing' },
-        }
-        const [ out ] = await registerServiceConsumerByTestClient(userClient, payload)
-        const [ meter ] = await Meter.getAll(userClient)
-
-        expect(out).toBeDefined()
-        expect(out.accountNumber).toEqual(payload.accountNumber)
-        expect(out.resident.id).toEqual(payload.residentId)
-        expect(out.organization.id).toEqual(payload.organizationId)
-        expect(out.paymentCategory).toEqual('Housing')
-        // TODO(zuch): Fix test
-        //expect(out.residentBillingAccount.id).toEqual(billingAccountAttrs.id)
-        //expect(out.residentOrganization.id).toEqual(userClient.organization.id)
-        //expect(out.residentAcquiringIntegrationContext.id).toEqual(acquiringIntegrationContext.id)
-        //expect(out.residentAcquiringIntegrationContext.integration).toEqual(acquiringIntegration.id)
-        expect(meter).toBeDefined()
-        expect(meter.number).toBeDefined()
-    })
-
-    it('creates serviceConsumer with billingAccount without Meters', async () => {
-        const userClient = await makeClientWithProperty()
-
-        const [context] = await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, integration, { status: CONTEXT_FINISHED_STATUS })
-
-        const [billingProperty] = await createTestBillingProperty(adminClient, context)
-        const [billingAccountAttrs] = await createTestBillingAccount(adminClient, context, billingProperty)
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: billingAccountAttrs.unitName,
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: billingAccountAttrs.number,
-            organizationId: userClient.organization.id,
-        }
-        const [ out ] = await registerServiceConsumerByTestClient(userClient, payload)
-
-        expect(out).toBeDefined()
-        expect(out.accountNumber).toEqual(payload.accountNumber)
-        expect(out.resident.id).toEqual(payload.residentId)
-        expect(out.organization.id).toEqual(payload.organizationId)
-    })
-
-    it('dont create serviceConsumer if billingIntegrationOrganizationContext is not finished', async () => {
-        const userClient = await makeClientWithProperty()
-        const adminClient = await makeLoggedInAdminClient()
-
-        const [integration] = await createTestBillingIntegration(adminClient)
-        const [context] = await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, integration)
-
-        const [billingProperty] = await createTestBillingProperty(adminClient, context)
-        const [billingAccountAttrs] = await createTestBillingAccount(adminClient, context, billingProperty)
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: billingAccountAttrs.unitName,
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: billingAccountAttrs.number,
-            organizationId: userClient.organization.id,
-        }
-
-        await expectToThrowGQLError(async () => {
-            await registerServiceConsumerByTestClient(userClient, payload)
-        }, {
-            mutation: 'registerServiceConsumer',
-            variable: ['data', 'accountNumber'],
-            code: 'BAD_USER_INPUT',
-            type: 'NOT_FOUND',
-            message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and organization combination',
-            messageForUser: 'api.resident.RegisterServiceConsumerService.BILLING_ACCOUNT_NOT_FOUND',
-        })
-    })
-
-    it('creates serviceConsumer without billingAccount when Meters are found', async () => {
-        const userClient = await makeClientWithProperty()
-
-        const [resource] = await MeterResource.getAll(adminClient, { id: COLD_WATER_METER_RESOURCE_ID })
-
-        const USER_UNIT_NAME = String(faker.datatype.number())
-        const USER_ACCOUNT_NUMBER = String(faker.datatype.number())
-
-        await createTestMeter(adminClient, userClient.organization, userClient.property, resource, {
-            unitName: USER_UNIT_NAME,
-            accountNumber: USER_ACCOUNT_NUMBER,
-        })
-        await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, integration)
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: USER_UNIT_NAME,
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: USER_ACCOUNT_NUMBER,
-            organizationId: userClient.organization.id,
-        }
-
-        const [ out ] = await registerServiceConsumerByTestClient(userClient, payload)
-        expect(out).toBeDefined()
-        expect(out.accountNumber).toEqual(payload.accountNumber)
-        expect(out.resident.id).toEqual(payload.residentId)
-        expect(out.organization.id).toEqual(payload.organizationId)
-        expect(out.residentBillingAccount).toBeNull()
-    })
-
-    it('fails with error when billingAccount not found, and Meters are not found', async () => {
-        const userClient = await makeClientWithProperty()
-
-        await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, integration, { status: CONTEXT_FINISHED_STATUS })
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: '21',
-        })
-
-        const payload = {
-            residentId: resident.id,
-            accountNumber: '221231232',
-            organizationId: userClient.organization.id,
-        }
-
-        await catchErrorFrom(async () => {
-            await registerServiceConsumerByTestClient(userClient, payload)
-        }, ({ errors }) => {
-            expect(errors).toMatchObject([{
-                message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and organization combination',
-                name: 'GQLError',
-                path: ['obj'],
-                extensions: {
-                    mutation: 'registerServiceConsumer',
-                    variable: ['data', 'accountNumber'],
-                    code: 'BAD_USER_INPUT',
-                    type: 'NOT_FOUND',
-                    message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and organization combination',
-                },
-            }])
-        })
-    })
-
-    it('fails with error when creating serviceConsumer for nullish data', async () => {
-        const userClient = await makeClientWithProperty()
-
-        const [context] = await createTestBillingIntegrationOrganizationContext(adminClient, userClient.organization, integration, { status: CONTEXT_FINISHED_STATUS })
-        const [billingProperty] = await createTestBillingProperty(adminClient, context)
-        const [billingAccountAttrs] = await createTestBillingAccount(adminClient, context, billingProperty)
-
-        await updateTestUser(adminClient, userClient.user.id, { type: RESIDENT })
-        const [resident] = await createTestResident(adminClient, userClient.user, userClient.property, {
-            unitName: billingAccountAttrs.unitName,
-        })
-
-        const payloadWithNullishAccountName = {
-            residentId: resident.id,
-            accountNumber: '',
-            organizationId: userClient.organization.id,
-        }
-
-        await catchErrorFrom(async () => {
-            await registerServiceConsumerByTestClient(userClient, payloadWithNullishAccountName)
-        }, ({ errors }) => {
-            expect(errors).toMatchObject([{
-                message: 'Argument "accountNumber" is null or empty',
-                name: 'GQLError',
-                path: ['obj'],
-                extensions: {
-                    mutation: 'registerServiceConsumer',
-                    variable: ['data', 'accountNumber'],
-                    code: 'BAD_USER_INPUT',
-                    type: 'WRONG_FORMAT',
-                    message: 'Argument "accountNumber" is null or empty',
-                },
-            }])
-        })
-    })
-
-    it('cannot be invoked by non-resident user', async () => {
-        const userClient = await makeClientWithProperty()
-
-        const payload = {
-            residentId: 'test-id',
-            accountNumber: 'test-number',
-            organizationId: userClient.organization.id,
-        }
-
-        await expectToThrowAccessDeniedErrorToObj(async () => {
-            await registerServiceConsumerByTestClient(userClient, payload)
-        })
-    })
-
-    it('cannot be invoked by anonymous', async () => {
-        const userClient = await makeClient()
-        const userClient2 = await makeClientWithProperty()
-
-        const payload = {
-            residentId: 'test-id',
-            accountNumber: 'test-number',
-            organizationId: userClient2.organization.id,
-        }
-
-        await expectToThrowAuthenticationErrorToObj(async () => {
-            await registerServiceConsumerByTestClient(userClient, payload)
-        })
-    })
-})
-
-describe('registerServiceConsumer use cases', () => {
+describe('RegisterServiceConsumer', () => {
 
     let utils
 
     beforeAll(async () => {
-        utils = new BillingTestUtils([AcquiringTestMixin, ResidentTestMixin])
+        utils = new BillingTestUtils([AcquiringTestMixin, PropertyTestMixin, ResidentTestMixin, MeterTestMixin])
         await utils.init()
-        await utils.initMixins()
     })
 
-    describe('Online interaction', () => {
+    describe('Common cases', () => {
+
+        test('allows to create service consumers with same resident and accountNumber for multiple organizations', async () => {
+            const resident = await utils.createResident()
+            const anotherUtils = new BillingTestUtils([AcquiringTestMixin])
+            await anotherUtils.init()
+            const accountNumber = faker.random.alphaNumeric(16)
+            await utils.createReceipts([ utils.createJSONReceipt({ accountNumber }) ])
+            await anotherUtils.createReceipts([ utils.createJSONReceipt({ accountNumber }) ])
+            const [consumer1] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
+                accountNumber,
+                organizationId: utils.organization.id,
+            })
+            const [consumer2] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
+                accountNumber,
+                organizationId: anotherUtils.organization.id,
+            })
+            expect(consumer1).toHaveProperty('id')
+            expect(consumer2).toHaveProperty('id')
+            expect(consumer1.id).not.toEqual(consumer2.id)
+        })
+
+        test('does not create same service consumer twice', async () => {
+            const resident = await utils.createResident()
+            const accountNumber = faker.random.alphaNumeric(16)
+            await utils.createReceipts([
+                utils.createJSONReceipt({ accountNumber }),
+            ])
+            const createInput = { residentId: resident.id, accountNumber, organizationId: utils.organization.id }
+            const [initialConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, createInput)
+            const [duplicatedConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, createInput)
+            expect(initialConsumer.id).toEqual(duplicatedConsumer.id)
+        })
+
+        test('cannot be invoked by anonymous', async () => {
+            const payload = { residentId: 'test-id', accountNumber: 'test-number', organizationId: utils.organization.id }
+            await expectToThrowAuthenticationErrorToObj(async () => {
+                await registerServiceConsumerByTestClient(utils.clients.anonymous, payload)
+            })
+        })
+
+        test('cannot be invoked by non-resident user', async () => {
+            const payload = { residentId: 'test-id', accountNumber: 'test-number', organizationId: utils.organization.id }
+            await expectToThrowAccessDeniedErrorToObj(async () => {
+                await registerServiceConsumerByTestClient(utils.clients.employee, payload)
+            })
+        })
+
+        test('fails with error when creating serviceConsumer for empty accountNumber', async () => {
+            const resident = await utils.createResident()
+            const accountNumber = ' '
+            await expectToThrowGQLError(async () => {
+                await registerServiceConsumerByTestClient(utils.clients.resident, {
+                    residentId: resident.id,
+                    accountNumber,
+                    organizationId: utils.organization.id,
+                })
+            }, ACCOUNT_NUMBER_IS_NOT_SPECIFIED)
+        })
+
+        test('fails with error when billingAccount not found, and Meters are not found', async () => {
+            const resident = await utils.createResident()
+            const notExistingAccountNumber = faker.random.alphaNumeric(16)
+            await expectToThrowGQLError(async () => {
+                await registerServiceConsumerByTestClient(utils.clients.resident, {
+                    residentId: resident.id,
+                    accountNumber: notExistingAccountNumber,
+                    organizationId: utils.organization.id,
+                })
+            }, BILLING_ACCOUNT_NOT_FOUND)
+        })
+
+        it('creates serviceConsumer with billingAccount and Meters', async () => {
+            const resident = await utils.createResident()
+            const accountNumber = faker.random.alphaNumeric(16)
+            await utils.createReceipts([
+                utils.createJSONReceipt({ accountNumber }),
+            ])
+            await utils.createMeter({ accountNumber })
+            const [serviceConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
+                accountNumber,
+                organizationId: utils.organization.id,
+            })
+            expect(serviceConsumer).toHaveProperty('id')
+        })
+
+    })
+
+    describe('BillingIntegration with online interaction', () => {
 
         const TEST_INTERACTION_URL = 'https://external-integration.com/api/check?token=some_securitey_token'
 
@@ -797,7 +143,7 @@ describe('registerServiceConsumer use cases', () => {
         })
 
         test('should throw an error if no billingAccount found and check for online interaction failed', async () => {
-            const residentClient = await utils.createResidentClient()
+            const resident = await utils.createResident()
             const notExistingAccountNumber = faker.random.alphaNumeric(16)
             const mockResponse = {
                 ok: true,
@@ -805,32 +151,25 @@ describe('registerServiceConsumer use cases', () => {
             }
             fetch.mockResolvedValueOnce(mockResponse)
             await expectToThrowGQLError(async () => {
-                await registerServiceConsumerByTestClient(residentClient, {
-                    residentId: residentClient.resident.id,
+                await registerServiceConsumerByTestClient(utils.clients.resident, {
+                    residentId: resident.id,
                     accountNumber: notExistingAccountNumber,
                     organizationId: utils.organization.id,
                 })
-            }, {
-                mutation: 'registerServiceConsumer',
-                variable: ['data', 'accountNumber'],
-                code: 'BAD_USER_INPUT',
-                type: 'NOT_FOUND',
-                message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and organization combination',
-                messageForUser: 'api.resident.RegisterServiceConsumerService.BILLING_ACCOUNT_NOT_FOUND',
-            })
+            }, BILLING_ACCOUNT_NOT_FOUND)
             expect(fetch).toHaveBeenCalledTimes(1)
         })
 
         test('should work if no billingAccount in condo, but check for online interaction succeeded', async () => {
-            const residentClient = await utils.createResidentClient()
+            const resident = await utils.createResident()
             const existingAccount = faker.random.alphaNumeric(16)
             const mockResponse = {
                 ok: true,
                 json: jest.fn().mockResolvedValue({ status: ONLINE_INTERACTION_CHECK_ACCOUNT_SUCCESS_STATUS }),
             }
             fetch.mockResolvedValueOnce(mockResponse)
-            const [newConsumer] = await registerServiceConsumerByTestClient(residentClient, {
-                residentId: residentClient.resident.id,
+            const [newConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
                 accountNumber: existingAccount,
                 organizationId: utils.organization.id,
             })
@@ -839,7 +178,7 @@ describe('registerServiceConsumer use cases', () => {
         })
 
         test('should not call online interaction if account already presence in condo', async () => {
-            const residentClient = await utils.createResidentClient()
+            const resident = await utils.createResident()
             const existingAccount = faker.random.alphaNumeric(16)
             await utils.createReceipts([
                 utils.createJSONReceipt({ accountNumber: existingAccount }),
@@ -849,8 +188,8 @@ describe('registerServiceConsumer use cases', () => {
                 json: jest.fn().mockResolvedValue({ status: ONLINE_INTERACTION_CHECK_ACCOUNT_SUCCESS_STATUS }),
             }
             fetch.mockResolvedValueOnce(mockResponse)
-            const [newConsumer] = await registerServiceConsumerByTestClient(residentClient, {
-                residentId: residentClient.resident.id,
+            const [newConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
                 accountNumber: existingAccount,
                 organizationId: utils.organization.id,
             })
@@ -867,101 +206,100 @@ describe('registerServiceConsumer use cases', () => {
         })
 
         test('should throw an error if billingContext is not in Finished status', async () => {
-            const residentClient = await utils.createResidentClient()
+            const resident = await utils.createResident()
             const accountNumberForSuccess = faker.random.alphaNumeric(16)
             const accountNumberForFailure = faker.random.alphaNumeric(16)
             await utils.createReceipts([
                 utils.createJSONReceipt({ accountNumber: accountNumberForSuccess }),
                 utils.createJSONReceipt({ accountNumber: accountNumberForFailure }),
             ])
-            const [successConsumer] = await registerServiceConsumerByTestClient(residentClient, {
-                residentId: residentClient.resident.id,
+            const [successConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
                 accountNumber: accountNumberForSuccess,
                 organizationId: utils.organization.id,
             })
             expect(successConsumer).toHaveProperty('id')
             await utils.updateBillingContext({ status: CONTEXT_IN_PROGRESS_STATUS })
             await expectToThrowGQLError(async () => {
-                await registerServiceConsumerByTestClient(residentClient, {
-                    residentId: residentClient.resident.id,
+                await registerServiceConsumerByTestClient(utils.clients.resident, {
+                    residentId: resident.id,
                     accountNumber: accountNumberForFailure,
                     organizationId: utils.organization.id,
                 })
-            }, {
-                mutation: 'registerServiceConsumer',
-                variable: ['data', 'accountNumber'],
-                code: 'BAD_USER_INPUT',
-                type: 'NOT_FOUND',
-                message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and organization combination',
-                messageForUser: 'api.resident.RegisterServiceConsumerService.BILLING_ACCOUNT_NOT_FOUND',
-            })
+            }, BILLING_ACCOUNT_NOT_FOUND)
         })
+
         test('should throw an error if acquiringContext is not in Finished status', async () => {
-            const residentClient = await utils.createResidentClient()
+            const resident = await utils.createResident()
             const accountNumberForSuccess = faker.random.alphaNumeric(16)
             const accountNumberForFailure = faker.random.alphaNumeric(16)
             await utils.createReceipts([
                 utils.createJSONReceipt({ accountNumber: accountNumberForSuccess }),
                 utils.createJSONReceipt({ accountNumber: accountNumberForFailure }),
             ])
-            const [successConsumer] = await registerServiceConsumerByTestClient(residentClient, {
-                residentId: residentClient.resident.id,
+            const [successConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
                 accountNumber: accountNumberForSuccess,
                 organizationId: utils.organization.id,
             })
             expect(successConsumer).toHaveProperty('id')
             await utils.updateAcquiringContext({ status: CONTEXT_IN_PROGRESS_STATUS })
             await expectToThrowGQLError(async () => {
-                await registerServiceConsumerByTestClient(residentClient, {
-                    residentId: residentClient.resident.id,
+                await registerServiceConsumerByTestClient(utils.clients.resident, {
+                    residentId: resident.id,
                     accountNumber: accountNumberForFailure,
                     organizationId: utils.organization.id,
                 })
-            }, {
-                mutation: 'registerServiceConsumer',
-                variable: ['data', 'accountNumber'],
-                code: 'BAD_USER_INPUT',
-                type: 'NOT_FOUND',
-                message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and organization combination',
-                messageForUser: 'api.resident.RegisterServiceConsumerService.BILLING_ACCOUNT_NOT_FOUND',
-            })
+            }, BILLING_ACCOUNT_NOT_FOUND)
         })
+
         test('should work if no billingContext in Finished status, but organization has Meter with the same accountNumber', async () => {
-
+            const resident = await utils.createResident()
+            await utils.updateBillingContext({ status: CONTEXT_IN_PROGRESS_STATUS })
+            const accountNumber = faker.random.alphaNumeric(16)
+            await utils.createMeter({ accountNumber })
+            const [successConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
+                accountNumber,
+                organizationId: utils.organization.id,
+            })
+            expect(successConsumer).toHaveProperty('id')
         })
-        test('should work if no acquiringContext in Finished status, but organization has Meter with the same accountNumber', async () => {
 
+        test('should work if no acquiringContext in Finished status, but organization has Meter with the same accountNumber', async () => {
+            const resident = await utils.createResident()
+            await utils.updateAcquiringContext({ status: CONTEXT_IN_PROGRESS_STATUS })
+            const accountNumber = faker.random.alphaNumeric(16)
+            await utils.createMeter({ accountNumber })
+            const [successConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
+                accountNumber,
+                organizationId: utils.organization.id,
+            })
+            expect(successConsumer).toHaveProperty('id')
         })
 
         test('should throw an error if no billingAccount found', async () => {
-            const residentClient = await utils.createResidentClient()
+            const resident = await utils.createResident()
             const accountNumberForSuccess = faker.random.alphaNumeric(16)
             const accountNumberForFailure = faker.random.alphaNumeric(16)
             await utils.createReceipts([
                 utils.createJSONReceipt({ accountNumber: accountNumberForSuccess }),
             ])
-            const [successConsumer] = await registerServiceConsumerByTestClient(residentClient, {
-                residentId: residentClient.resident.id,
+            const [successConsumer] = await registerServiceConsumerByTestClient(utils.clients.resident, {
+                residentId: resident.id,
                 accountNumber: accountNumberForSuccess,
                 organizationId: utils.organization.id,
             })
             expect(successConsumer).toHaveProperty('id')
             await expectToThrowGQLError(async () => {
-                await registerServiceConsumerByTestClient(residentClient, {
-                    residentId: residentClient.resident.id,
+                await registerServiceConsumerByTestClient(utils.clients.resident, {
+                    residentId: resident.id,
                     accountNumber: accountNumberForFailure,
                     organizationId: utils.organization.id,
                 })
-            }, {
-                mutation: 'registerServiceConsumer',
-                variable: ['data', 'accountNumber'],
-                code: 'BAD_USER_INPUT',
-                type: 'NOT_FOUND',
-                message: 'Can\'t find billingAccount and any meters with this accountNumber, unitName and organization combination',
-                messageForUser: 'api.resident.RegisterServiceConsumerService.BILLING_ACCOUNT_NOT_FOUND',
-            })
+            }, BILLING_ACCOUNT_NOT_FOUND)
         })
-
     })
 
 })
