+++ apps/condo/domains/resident/utils/testSchema/residentTestMixin.js
@@ -1,20 +1,18 @@
 const { createTestResident } = require('./index')
-const { createTestProperty } = require("@condo/domains/property/utils/testSchema");
 const { faker} = require("@faker-js/faker");
 const { makeClientWithResidentUser } = require("@condo/domains/user/utils/testSchema");
 
 const ResidentTestMixin = {
 
-    async initMixin () {},
-    async createResidentClient () {
-        const [property] = await createTestProperty(this.clients.admin, this.organization)
-        const unitName = faker.random.alphaNumeric(8)
-        const residentClient = await makeClientWithResidentUser()
-        const [resident] = await createTestResident(this.clients.admin, residentClient.user, property, {
-            unitName,
+    async initMixin () {
+        this.clients.resident = await makeClientWithResidentUser()
+    },
+    async createResident ({ unitName, unitType } = {}) {
+        const [resident] = await createTestResident(this.clients.admin, this.clients.resident.user, this.property, {
+            unitName: unitName || faker.random.alphaNumeric(8),
+            unitType: unitType || 'flat',
         })
-        residentClient.resident = resident
-        return residentClient
+        return resident
     }
 
 }
